pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contentflow-api"
version = "0.2.0"
description = "Social media distribution and AI video generation API."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.40.0,<1.0.0",
  "bcrypt>=4.1.2,<5.0.0",
  "celery>=5.4.0,<6.0.0",
  "cryptography>=43.0.0,<44.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "PyJWT>=2.8.0,<3.0.0",
  "redis[hiredis]>=5.0.8,<6.0.0",
  "sentry-sdk[fastapi]>=2.14.0,<3.0.0",
  "stripe>=11.6.0,<12.0.0",
  "structlog>=24.4.0,<25.0.0",
  "supabase>=2.6.0,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.2,<2.0.0",
  "fakeredis>=2.24.0,<3.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "respx>=0.21.0,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[tool.setuptools]
packages = ["app", "app.adapters", "app.api", "app.api.v1", "app.core", "app.middleware", "app.models", "app.oauth", "app.oauth.providers", "app.services", "app.workers"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = true
